Crate tx_bakery

Transaction Bakery

Modules

chain_query

Trait for a Cardano chain query client

error

metadata

Transaction Metadata

submitter

Trait for a component capable of submitting transactions

time

POSIX time to/from slot conversions

tx_info_builder

Transaction Info builder

utils

wallet

Wallet trait

Structs

TxBakery

Transaction builder

TxWithCtx

TransactionInfo with additional context required to build a valid transaction

Enums

ChangeStrategy

Options to deal with change outputs and collateral returns

CollateralStrategy

Options to deal with collateral selection