tx_bakery
Module chain_query

Trait for a Cardano chain query client

Structs

ChainQueryError

EraParameters

EraSummary

EraTime

FullTransactionOutput

ProtocolParameters

A subset of Cardano protocol parameters, only handling values that we use for transaction building

Enums

ChainTip

Network

Cardano network discriminant

Traits

ChainQuery

A chain query client responsible for all read actions from the blockchain (no write)