tx_bakery::wallet
Trait Wallet

pub trait Wallet {
    // Required methods
    fn sign_transaction(&self, tx: &FixedTransaction) -> FixedTransaction;
    fn get_change_pkh(&self) -> Ed25519PubKeyHash;
    fn get_change_addr(&self) -> Address;
}

Cardano wallet that has access to a private key (directly or indirectly) and able to sign a transaction

Required Methods

fn sign_transaction(&self, tx: &FixedTransaction) -> FixedTransaction

Signs a fully built transaction

fn get_change_pkh(&self) -> Ed25519PubKeyHash

Query the public key hash used by this wallet

fn get_change_addr(&self) -> Address

Query the wallet address